+++ test/barcodes/test_barcodes.py
@@ -9,9 +9,9 @@ HERE = Path(__file__).resolve().parent
 def test_code39(tmp_path):
     pdf = FPDF()
     pdf.add_page()
-    pdf.code39("fpdf2", x=50, y=50, w=4, h=20)
+    pdf.code39("*fpdf2*", x=30, y=50, w=4, h=20)
     pdf.set_font("courier", "B", size=36)
-    pdf.text(x=80, y=80, txt="fpdf2")
+    pdf.text(x=70, y=80, txt="*fpdf2*")
     assert_pdf_equal(pdf, HERE / "barcodes_code39.pdf", tmp_path)